move footer logic
